Make a helper loader
